Move ol.tilegrid.XYZ tests to their proper place
